Move RideFilter to :drive module from :core:entities
